hybrid_rag/requirements.txt
# Core
openai>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Vector store
chromadb>=0.4.0

# Graph (LlamaIndex)
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-llms-openai-like>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-embeddings-openai-like>=0.1.0
networkx>=3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# UI
streamlit>=1.30.0
requests>=2.28.0
